pyproject.toml
[project]
name = "ton-split-ai"
version = "0.1.0"
description = "Telegram-native AI receipt splitting bot with TON payment links"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.20.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.35",
  "aiosqlite>=0.20.0",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "openai>=1.68.0",
  "pillow>=10.4.0",
  "opencv-python-headless>=4.10.0.84",
  "python-multipart>=0.0.9",
  "orjson>=3.10.7",
  "loguru>=0.7.2",
  "python-dotenv>=1.0.1",
  "qrcode[pil]>=7.4.2",
  "httpx>=0.27.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
